+++ Screenshot/src/Screenshot.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26228.4
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Screenshot", "Screenshot.vcxproj", "{725876EB-B83D-4F42-9988-BB9F2DD41039}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|win32 = Debug|win32
+		Debug|x64 = Debug|x64
+		Release|win32 = Release|win32
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{725876EB-B83D-4F42-9988-BB9F2DD41039}.Debug|win32.ActiveCfg = Debug|Win32
+		{725876EB-B83D-4F42-9988-BB9F2DD41039}.Debug|win32.Build.0 = Debug|Win32
+		{725876EB-B83D-4F42-9988-BB9F2DD41039}.Debug|x64.ActiveCfg = Debug|x64
+		{725876EB-B83D-4F42-9988-BB9F2DD41039}.Debug|x64.Build.0 = Debug|x64
+		{725876EB-B83D-4F42-9988-BB9F2DD41039}.Release|win32.ActiveCfg = Release|Win32
+		{725876EB-B83D-4F42-9988-BB9F2DD41039}.Release|win32.Build.0 = Release|Win32
+		{725876EB-B83D-4F42-9988-BB9F2DD41039}.Release|x64.ActiveCfg = Release|x64
+		{725876EB-B83D-4F42-9988-BB9F2DD41039}.Release|x64.Build.0 = Release|x64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
