--- ScreenshotReciever/packages.config
@@ -1,13 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="sfml-audio" version="2.4.0.0" targetFramework="native" />
-  <package id="sfml-audio.redist" version="2.4.0.0" targetFramework="native" />
-  <package id="sfml-graphics" version="2.4.0.0" targetFramework="native" />
-  <package id="sfml-graphics.redist" version="2.4.0.0" targetFramework="native" />
-  <package id="sfml-network" version="2.4.0.0" targetFramework="native" />
-  <package id="sfml-network.redist" version="2.4.0.0" targetFramework="native" />
-  <package id="sfml-system" version="2.4.0.0" targetFramework="native" />
-  <package id="sfml-system.redist" version="2.4.0.0" targetFramework="native" />
-  <package id="sfml-window" version="2.4.0.0" targetFramework="native" />
-  <package id="sfml-window.redist" version="2.4.0.0" targetFramework="native" />
-</packages>
