+++ ScreenshotReciever/src/ScreenshotReciever.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26228.4
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ScreenshotReciever", "ScreenshotReciever.vcxproj", "{3F951250-B645-4611-8EE0-34F331F35F0D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Win32 = Debug|Win32
+		Debug|x64 = Debug|x64
+		Release|Win32 = Release|Win32
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3F951250-B645-4611-8EE0-34F331F35F0D}.Debug|Win32.ActiveCfg = Debug|Win32
+		{3F951250-B645-4611-8EE0-34F331F35F0D}.Debug|Win32.Build.0 = Debug|Win32
+		{3F951250-B645-4611-8EE0-34F331F35F0D}.Debug|x64.ActiveCfg = Debug|x64
+		{3F951250-B645-4611-8EE0-34F331F35F0D}.Debug|x64.Build.0 = Debug|x64
+		{3F951250-B645-4611-8EE0-34F331F35F0D}.Release|Win32.ActiveCfg = Release|Win32
+		{3F951250-B645-4611-8EE0-34F331F35F0D}.Release|Win32.Build.0 = Release|Win32
+		{3F951250-B645-4611-8EE0-34F331F35F0D}.Release|x64.ActiveCfg = Release|x64
+		{3F951250-B645-4611-8EE0-34F331F35F0D}.Release|x64.Build.0 = Release|x64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
