+++ ScreenshotReciever/src/packages.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="sfml-network" version="2.4.0.0" targetFramework="native" />
+  <package id="sfml-network.redist" version="2.4.0.0" targetFramework="native" />
+  <package id="sfml-system" version="2.4.0.0" targetFramework="native" />
+  <package id="sfml-system.redist" version="2.4.0.0" targetFramework="native" />
+</packages>
